Close #12 - Rename package from kevinlee.testosterone to testosterone
