--- src/main/java/kevinlee/testosterone/TestResultHandler.java
@@ -1,30 +1,0 @@
-/**
- * Copyright 2014 Seong Hyun (Kevin)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package kevinlee.testosterone;
-
-/**
- * @author Lee, SeongHyun (Kevin)
- * @version 0.0.1 (2014-08-17)
- *
- */
-public interface TestResultHandler<T> {
-
-  String getTestInfo();
-
-  <EX extends Throwable> void expect(final ExpectedExceptionAssertions<EX> expectedExceptionAssertion);
-
-  Then<T> then(final T then);
-}
