+++ src/main/java/testosterone/CallableTestResultHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kevinlee.testosterone;
+package testosterone;
 
 import java.util.concurrent.Callable;
 import java.util.function.Consumer;
