+++ src/main/java/testosterone/ExpectedExceptionAssertions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kevinlee.testosterone;
+package testosterone;
 
 import java.util.function.Function;
