+++ src/main/java/testosterone/RunnableTestResultHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kevinlee.testosterone;
+package testosterone;
 
 /**
  * @author Lee, SeongHyun (Kevin)
